fix build errors in loopback_test_nf7
The /lib/hw/xilinx pcore folder contains IP that isn't supported for the
kintex7 architecture. Rather than patch the files in the xilinx lib,
there are now symbolic links to the contrib and std libraries just for
this project.

This may not work on operating systems that do not support symbolic
links.

Also, updated the bitfile and added an ignore file to the embedded sw
folder.