Made to be more like the orig reference_flash project
Want to minimize confusion about use of this tool compared to original
reference_flash tool These changes help minimize the differences.

The overall appearance of the project (filenames, etc) have been changed
to be more consistent with the overall repository scheme. Debug programs
have been removed.

The host executable has been both re-named and argument handling have
been updated in an effort to be closer to the original in use.